HEALTHAI/healthai/requirements.txt
streamlit
pandas
numpy
plotly
requests
python-dotenv